// src/app/models/renfacfo.ts
export interface Renfacfo {
    idrenfacfo: number;
    idcli: number;
    idfac: number;
    codigo: string;
    concepto: string;
    serie: string;
    canti: number;
    preciou: number;
    piva: number;
    pdsc: number;
    importe: number;
    descto: number;
    iva: number;
    cia: number;
    folio: number;
}